Support icon packs,per-app icon customization, app label customization
---
ezio84
- port to our Oreo Pixel Launcher mod
- fix blank icon for Oreo AdaptiveIconDrawable icons
- move settings to our fragments and shortcuts popups
- dinamycally theme icon edit dialogs and popups
- todo: when using a custom icon pack that doesn't have an icon for an
app with AdaptiveIconDrawable icon, the square shape is set for the
background drawable if the icon pack doesn't have an iconback drawable.
It'd be good instead to mask the background to the shape enabled by the
user. See IconsHandler.java getDefaultAppDrawable, we could mask
backgroundDr to:
AdaptiveIconDrawable backgroundDr = (AdaptiveIconDrawable)
   mContext.getDrawable(R.drawable.adaptive_icon_drawable_wrapper).mutate();
the same issue happens when resetting the icon if using the stock icon
pack, that's why atm i disabled icon customization for it
- todo2: cleanup and improve stuff
---

This introduces a new tab called customization.
It can be accessed by longpressing the homescreen and pressing the paint
icon.
It supports selecting a custom icon pack.
Once a custom icon pack has been selected you can drag any icon you want
to the edit drop target.
Once done, it will launch a dialog allowing you to change the app name.
When pressing on the app icon inside the dialog, an activity will launch
giving you the option to select a new icon for your current app.

Ticket: NOUGAT-56

Change-Id: I4a55dfb2ca42be20d2e81e7f8ec2a8ae207ee165

Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>

Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>